pip._vendor: do not bother try/except only to immediately and always reraise
The original intent of this code went through numerous different
iterations, mainly centered around the idea that ImportErrors should be
squelched (but only in the initial import, not the success branch in the
else clause, as per commit 5a94590f0e261cda2e166710855f14e3b2296676.

Most recently, this entire approach was recanted and debundled modules
are required to exist and must import up-front. But rather than
reverting to commit 5a94590f0e261cda2e166710855f14e3b2296676, the
implementation detail of using a try/except was retained... even though
there is nothing left to catch, the ImportError is immediately reraised
without any modifications.

Either a try/except: raise, or a simple import, is equally, immediately,
fatal. But, the second version is fewer lines of code. So use that
instead.